Move some code over the mem overrides
