+++ src/enclave/enclave_mem.c
@@ -26,8 +26,6 @@ static void* mmap_base;         // First page that can be mmap'ed
 static void* mmap_end;          // Last page that can be mmap'ed
 static size_t mmap_num_pages;   // Total number of pages that can be mmap'ed
 
-static int mmap_files; // Allow MAP_PRIVATE or MAP_SHARED?
-
 static size_t used_pages =
     0; // Tracks the number of used pages for the mmap tracing
 
@@ -94,57 +92,6 @@ static void* index_to_addr(size_t index)
 static size_t addr_to_index(void* addr)
 {
     return ((char*)mmap_end - (char*)addr) / PAGE_SIZE;
-}
-
-long syscall_SYS_mmap(
-    void* addr,
-    size_t length,
-    int prot,
-    int flags,
-    int fd,
-    off_t offset)
-{
-    if ((flags & MAP_SHARED) && (flags & MAP_PRIVATE))
-    {
-        sgxlkl_warn("mmap() with MAP_SHARED and MAP_PRIVATE not supported\n");
-        return -EINVAL;
-    }
-    // Anonymous mapping/allocation
-    else if (fd == -1 && (flags & MAP_ANONYMOUS))
-    {
-        return (long)enclave_mmap(addr, length, flags & MAP_FIXED, prot, 1);
-    }
-    // File-backed mapping (if allowed)
-    else if ((fd >= 0) && enclave_mmap_files_flags_supported(flags))
-    {
-        void* mem =
-            enclave_mmap(addr, length, flags & MAP_FIXED, prot | PROT_WRITE, 0);
-
-        if (mem > 0)
-        {
-            // TODO:
-            // handle reading less than length
-            // handle EOF before length is read
-            ssize_t r = pread(fd, mem, length, offset);
-
-            if (r < 0)
-            {
-                oe_host_printf("A3: returning\n");
-                enclave_munmap(addr, length);
-                return -EACCES;
-            }
-
-            // Set requested page permissions
-            if ((prot | PROT_WRITE) != prot)
-                mprotect(mem, length, prot);
-        }
-
-        return (long)mem;
-    }
-    else
-    {
-        return -EINVAL;
-    }
 }
 
 void* syscall_SYS_mremap(
@@ -231,22 +178,6 @@ void enclave_mman_init(const void* base, size_t num_pages, int _mmap_files)
     bitmap_set(mmap_fresh_bitmap, 0, mmap_num_pages);
 
     mmap_files = _mmap_files;
-}
-
-/*
- * Returns 1 if we can mmap files using the given flags
- * returns 0 otherwise.
- */
-int enclave_mmap_files_flags_supported(int flags)
-{
-    int supported_flags = 0;
-
-    if (mmap_files == ENCLAVE_MMAP_FILES_SHARED)
-        supported_flags = MAP_PRIVATE | MAP_SHARED;
-    else if (mmap_files == ENCLAVE_MMAP_FILES_PRIVATE)
-        supported_flags = MAP_PRIVATE;
-
-    return supported_flags & flags;
 }
 
 /*
