+++ src/include/lkl/syscall-overrides-mem.h
@@ -1,0 +1,9 @@
+#ifndef _LKL_SYSCALL_OVERRIDES_MEM_H
+#define _LKL_SYSCALL_OVERRIDES_MEM_H
+
+/**
+ * Register override functions for the memory management functions.
+ */
+void syscall_register_mem_overrides();
+
+#endif
